requirements.txt
langgraph>=0.2
langchain-aws>=0.2
langchain-core>=0.3
python-dotenv>=1.0
boto3>=1.34
streamlit>=1.35
langchain-mcp-adapters>=0.1
langchain